requirements.txt
requests>=2.28.0,<3
python-dotenv>=0.19.0
pytest>=7.0.0,<8
